requirements.txt
requests==2.32.3
PyYAML==6.0.1
pymodbus==3.6.2
python-dotenv==1.0.1